Changed Light Flickering Script to Sync with Round Manager Script
Modified the Light Flickering Script and synced it with the Round Manager Script. Had to change up the AnomalyTimerUI Script too in order to be able to see the anomaly on the anomaly timer panel and also added the new anomaly type into the enums of the AnomalyContracts Script. Lastly implemented the Light Flicker anomaly into all the lights within the house.